requirements.txt
# requirements.txt
# Dépendances du backend AI Scrum PO Assistant
# Pour installer : pip install -r requirements.txt


annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.121.1
git-filter-repo==2.47.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.11.1
numpy==2.3.4
openai==2.7.1
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wavio==0.0.9
